// sdk/core/core-websockets/src/models.ts
export interface CloseInfo {
  /**
   * The close code.
   */
  code?: string;
  /**
   * The close reason.
   */
  reason?: string;
}

/**
 * Options to close the connection.
 */
export interface CloseOptions {
  /**
   * The close info.
   */
  info?: CloseInfo;
  /**
   * The abort signal.
   */
  abortSignal?: AbortSignal;
}

/**
 * Options to open the connection.
 */
export interface OpenOptions {
  /**
   * The abort signal.
   */
  abortSignal?: AbortSignal;
}

/**
 * Options to send data.
 */
export interface SendOptions {
  /**
   * The abort signal.
   */
  abortSignal?: AbortSignal;
}
// ...
export type Data = string | ArrayBuffer | ArrayBufferView;

/**
 * The WebSocket client.
 */
export interface WebSocketClient {
  /**
   * Send a message to the WebSocket server.
   * @param data - The message to send.
   *
   * @returns false if the send buffer is full
   */
  send(data: Data, opts?: SendOptions): Promise<boolean>;

  /** Registers a callback to be called when a message is received */
  on(event: "message", listener: (data: Data) => void): void;

  /** Registers a callback to be called when an open event is received */
  on(event: "reconnect", listener: () => void): void;

  /** Registers a callback to be called when a close event is received */
  on(event: "close", listener: () => void): void;

  /** Registers a callback to be called when an error event is received */
  on(event: "error", listener: (error: Error) => void): void;

  /**
   * Close the WebSocket connection.
   */
  close(opts?: CloseOptions): Promise<void>;
}
// ...
export interface WebSocketClientOptions {
  /**
   * The WebSocket protocol version(s) to use.
   * This can be a single protocol string or an array of protocol strings.
   */
  protocols?: string | string[];

  /**
   * NODEJS ONLY and WS ONLY
   *
   * The options to create the WS client.
   */
  wsOptions?: WS.ClientOptions | ClientRequestArgs;

  /**
   * The high water mark for the send queue.
   */
  highWaterMark?: number;

  /**
   * Whether to allow insecure connections.
   */
  allowInsecureConnection?: boolean;

  /**
   * Options for retrying the connection in case of failure.
   */
  retryOptions?: RetryOptions;

  /**
   * A unique identifier for the WebSocket client.
   */
  identifier?: string;
}